Date/Time threshold initial trial

+++ src/zencharts/data/DatePoint.java
@@ -5,23 +5,23 @@ import java.util.Date;
 
 import android.text.format.DateFormat;
 
-public class DatePoint implements Serializable{
+public class DatePoint implements Serializable {
 
 	/**
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
 	public long timeStamp;
-	public int value;
+	public float value;
 	public String label = "";
 	public String dateString = "";
-	
-	public DatePoint(long inTimestamp, int inValue, String inlabel) {
+
+	public DatePoint(long inTimestamp, float inValue, String inlabel) {
 		super();
 		timeStamp = inTimestamp;
 		value = inValue;
 		label = inlabel;
-		
+
 		Date date = new Date();
 		date.setTime(timeStamp);
 		dateString = DateFormat.format("MM/dd/yy", date).toString();
